Upgrade to dotnet 7
