+++ Test/Test.csproj
@@ -1,0 +1,25 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+
+        <IsPackable>false</IsPackable>
+
+        <Nullable>enable</Nullable>
+
+        <Platforms>AnyCPU;x64</Platforms>
+
+        <TargetFramework>net7.0-windows</TargetFramework>
+
+        <LangVersion>default</LangVersion>
+
+        <RootNamespace>testefiling</RootNamespace>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="nunit" Version="3.13.3"/>
+        <PackageReference Include="NUnit3TestAdapter" Version="4.3.1"/>
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.1"/>
+        <ProjectReference Include="..\efiling\efiling.csproj"/>
+    </ItemGroup>
+
+</Project>
