--- efiling.sln
@@ -1,41 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31019.35
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "efiling", "efiling\efiling.csproj", "{D2E2300A-4C49-4063-9088-2AA88AFB4D68}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "testefiling", "testefiling\testefiling.csproj", "{816139C4-9950-425C-AF66-1A7C9E06828D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D2E2300A-4C49-4063-9088-2AA88AFB4D68}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D2E2300A-4C49-4063-9088-2AA88AFB4D68}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D2E2300A-4C49-4063-9088-2AA88AFB4D68}.Debug|x64.ActiveCfg = Debug|x64
-		{D2E2300A-4C49-4063-9088-2AA88AFB4D68}.Debug|x64.Build.0 = Debug|x64
-		{D2E2300A-4C49-4063-9088-2AA88AFB4D68}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D2E2300A-4C49-4063-9088-2AA88AFB4D68}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D2E2300A-4C49-4063-9088-2AA88AFB4D68}.Release|x64.ActiveCfg = Release|x64
-		{D2E2300A-4C49-4063-9088-2AA88AFB4D68}.Release|x64.Build.0 = Release|x64
-		{816139C4-9950-425C-AF66-1A7C9E06828D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{816139C4-9950-425C-AF66-1A7C9E06828D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{816139C4-9950-425C-AF66-1A7C9E06828D}.Debug|x64.ActiveCfg = Debug|x64
-		{816139C4-9950-425C-AF66-1A7C9E06828D}.Debug|x64.Build.0 = Debug|x64
-		{816139C4-9950-425C-AF66-1A7C9E06828D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{816139C4-9950-425C-AF66-1A7C9E06828D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{816139C4-9950-425C-AF66-1A7C9E06828D}.Release|x64.ActiveCfg = Release|x64
-		{816139C4-9950-425C-AF66-1A7C9E06828D}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CEB597C2-7279-45CD-BF24-4381197BFAF2}
-	EndGlobalSection
-EndGlobal
