+++ efiling.sln.DotSettings.user
@@ -4,6 +4,7 @@
     &lt;TestId&gt;NUnit3x::816139C4-9950-425C-AF66-1A7C9E06828D::net5.0-windows7.0::testefiling.Tests.generatePdf&lt;/TestId&gt;&#xD;
   &lt;/TestAncestor&gt;&#xD;
 &lt;/SessionState&gt;</s:String>
+	<s:Boolean x:Key="/Default/Environment/WinFormsDesigner/IsDesignerExceptionPopupActive/@EntryValue">True</s:Boolean>
 	<s:Boolean x:Key="/Default/ResxEditorPersonal/CheckedGroups/=efiling_002FProperties_002FResources/@EntryIndexedValue">True</s:Boolean>
