+++ efiling/App.xaml
@@ -1,4 +1,4 @@
-﻿<Application x:Class="efiling.App"
+﻿<Application x:Class="EFiling.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              StartupUri="MainWindow.xaml">
