+++ efiling/App.xaml.cs
@@ -1,12 +1,12 @@
 ﻿using System.Windows;
 
-namespace efiling {
-    /// <summary>
-    ///     Interaction logic for App.xaml
-    /// </summary>
-    public partial class App : Application {
-        public App() {
-            InitializeComponent();
-        }
+namespace EFiling;
+
+/// <summary>
+///     Interaction logic for App.xaml
+/// </summary>
+public partial class App : Application {
+    public App() {
+        InitializeComponent();
     }
 }
