+++ efiling/model/Data.cs
@@ -1,13 +1,13 @@
-﻿namespace efiling.model {
-    public class Data {
-        public Data(Person advocate, Person respondent, CaseDetails caseDetails) {
-            Advocate = advocate;
-            Respondent = respondent;
-            CaseDetails = caseDetails;
-        }
+﻿namespace EFiling.Model;
 
-        public Person Advocate { get; }
-        public Person Respondent { get; }
-        public CaseDetails CaseDetails { get; }
+public class Data {
+    public Data(Person advocate, Person respondent, CaseDetails caseDetails) {
+        Advocate = advocate;
+        Respondent = respondent;
+        CaseDetails = caseDetails;
     }
+
+    public Person Advocate { get; }
+    public Person Respondent { get; }
+    public CaseDetails CaseDetails { get; }
 }
