--- testefiling/testefiling.csproj
@@ -1,21 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-
-        <IsPackable>false</IsPackable>
-
-        <Nullable>enable</Nullable>
-
-        <TargetFramework>net5.0-windows</TargetFramework>
-
-        <Platforms>AnyCPU;x64</Platforms>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="nunit" Version="3.13.1" />
-        <PackageReference Include="NUnit3TestAdapter" Version="3.17.0" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.1" />
-        <ProjectReference Include="..\efiling\efiling.csproj"/>
-    </ItemGroup>
-
-</Project>
